requirements.txt
requests
beautifulsoup4
python-docx
groq
pypdf
streamlit
python-pptx
